Removing remaining calls to subprocess and replacing with score_db imports

+++ scripts/db_inc_logs.py
@@ -23,7 +23,9 @@ import os
 import pathlib
 import datetime as dt
 from dotenv import load_dotenv
-import subprocess
+
+from score_db import score_db_base
+from score_db import file_utils
 
 #DICTIONARIES
 #file list needed to harvest
@@ -50,7 +52,7 @@ datetime_str = datetime_obj.strftime("%Y%m%d%H")
 cycle_str = datetime_obj.strftime("%Y-%m-%d %H:%M:%S")
 
 input_env = sys.argv[2]
-env_path = os.path.join(pathlib.Path(__file__).parent.resolve(), input_env)
+env_path = os.path.join(pathlib.Path(__file__).parent.parent.resolve(), input_env)
 load_dotenv(env_path)
 
 s3 = boto3.resource(
@@ -94,8 +96,16 @@ for file in file_list:
     }
     yaml_file = db_yaml_generator.generate_harvest_metrics_yaml(os.getenv('EXPERIMENT_NAME'), os.getenv('EXPERIMENT_WALLCLOCK_START'),
                                                     'inc_logs', harvest_config)
+    
+    # validate configuration (yaml) file
+    file_utils.is_valid_readable_file(yaml_file)
+    # submit the score db request    
     print("Calling score-db with yaml file: " + yaml_file + "for cycle: " + cycle_str)
-    subprocess.run(["python3", os.getenv("SCORE_DB_BASE_LOCATION"), yaml_file], check=True)
+    response = score_db_base.handle_request(yaml_file)
+    if not response.success:
+        print(response.message)
+        print(response.errors)
+        raise RuntimeError("score-db returned a failure message") #generic exception to tell cylc to stop running 
 
     #remove yaml and downloaded file 
     os.remove(yaml_file)
